Bernstein-Yang: minor refactoring
Extracts `from_uint` and `to_uint` methods onto `Uint62L`, rather than
the previous `sat_to_unsat` and `unsat_to_sat` free functions.

Also extracts the `impl_limb_convert!` macro into its own module, which
will assist impl'ing a `BoxedUint62L`.
